refactor: components path to it's own dir Fixes #61
